nlp-service/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
Werkzeug==3.0.1
requests==2.31.0
pymongo==4.6.0
python-dotenv==1.0.0
gunicorn==21.2.0
